Add Pots of Gold problem and (Python) solution
This problem (and maybe the k egg-drop problem) are my favorite dynamic
programming problems. This is fairly simple, but I was surprised by the
non-intuitive play of the algorithm. At one point, I had an extra else
conditional in there to handle the case where `leftMove > rightMove`
and `leftMove + afterLeftMove == rightMove + afterRightMove`, but it
didn’t make the play any more intuitive.