Add isBST problem and solutions
I’m not sure why this problem is so popular. Maybe because the “binary
search” aspect of BSTs has a nice feel to it and everybody likes
testing recursion. There is also a “simple but wrong” recursive answer
that might catch people (where you don’t pass down the min/max).