Add Valid Interleaving problem
Probably one of my favorite questions. It runs in O(n) time and takes
O(n) space (though that space complexity is really in only the
worst-case of two identical strings). There is the finesse aspect of
having to add the duplicates back to the string it didn’t come from
that I like.