Add Linked List Cycle Problem
This was the first ‘practice interview question’ that I did back in the
day. I remember thinking that it was a ‘trick question’ because you
either knew the answer, or you didn’t. It demanded pure insight and
didn’t seem to have a recursive answer (which is what I expect from
linked lists). There is the space inefficient answer of either keeping
flags on the nodes themselves or keeping the memory addresses of the
nodes already seen. If you see enough of these types of questions, you
can start to learn the ‘insight’ to get new ‘trick’ questions, but I’m
not sure whether that is the point or not. The purpose of interviewing
shouldn’t be to test how good you are at interviewing. However, if the
insight you learn from preparing for interviews helps you with
real-life programming, then why not encourage it?