Add Median of Sorted Lists Problem
What a bi***. I was given this question in an interview once and pretty
quickly got the recursive answer to the equal-lengths version. But then
they asked for unequal lengths. I got the (recursive call) for
eliminating the floor(N/2) from both arrays but spent the remaining 8
minutes struggling with the base cases. How do you take the median of a
sorted array of 2 elements and a sorted array of 4 elements? It’s not
immediately evident that you have to use all 2+4=6 elements, but you do
because it could look like [1, 2], [3, 4, 5, 6] => 3.5. The use of max
and min functions hides the true number of comparisons that you have to
do. And you can’t treat M=2 like you treat an even M above 2. So while
enumerating the number of base cases, it seems to go on and on and on.
I know the purpose of interviews is seeing how you approach a problem
and not whether you finish solving it, but in reality, working on a
problem by approaching it in a methodical way that will take you more
than the allotted time is frustrating.