Add Next Greater Number
I like this question. It’s a basic “use a stack” question. Also it’s
O(n) time and space, but has strict guarantees on the number of
comparisons (the minimum being for a sorted array, which uses all of n
space in the stack but has the minimum of n comparisons; the maximum
being for a reversed sorted list with the largest number appended: [4,
3, 2, 1, 5] (because we check to see if the stack exists anyway), which
does 2n -1 comparisons). There is a similar question
(http://cs.stackexchange.com/questions/6898/finding-number-of-smaller-el
ements-for-each-element-in-an-array-efficiently) that I’ve been unable
to find an O(n) solution to and it keeps me up at night whenever I
think about it.