Fix find_median when lower is greater than upper
Evidently, it would fail on input [3, 4], [1, 2, 2] because `lower` would be the `max(3, 1) = 3` and `upper` would be `min(4, 2) = 2` so `lower > upper` which messed things up. That is, `median2 = 2` would be compared to `lower = 3` and since it was expected `median2 < lower < upper`, it would return `lower` as the answer, when really `median2 < upper < lower`. The quick fix is to switch `lower` and `upper` when `lower > upper`