My valid_interleaving solutions are wrong
I was very proud of my O(n) solution with O(n) space, and then this silly test case ruined it: is_valid_interleaving("aa","ab","aaba"). What my algorithm does is adds 'a' to duplicates, and then, upon seeing 'a' in str1 and not in str2, it assumes str1 was the correct way to interleave and continues on its merry way. It should also be looking in duplicates, but if it finds it in str1 (not in str2) and in duplicates, what should it do?